seperating Array generator
