+++ pdks/FinFET14nm_Mock_PDK/__init__.py
@@ -1,5 +1,6 @@
 from .cap import CapGenerator
 from .res import ResGenerator
+from .mos_array import MOSArrayGenerator
 from .mos import MOSGenerator
 from .guard_ring import RingGenerator
 from align.primitive.default.via import ViaArrayGenerator
